Move widgets into seperate folder
