+++ src/components/widgets/__test__/Loading.test.js
@@ -8,16 +8,8 @@ describe('Loading', () => {
     render(<Loading />);
 
     const loadingEl = await screen.findByTestId('loading');
-    const spinner = screen.getByRole('status');
 
     expect(loadingEl).toMatchSnapshot();
-    expect(spinner).toBeInTheDocument();
     expect(screen.getByTestId('loading')).toBeInTheDocument();
-  });
-
-  it('should render with given variant', () => {
-    render(<Loading variant="sm" />);
-
-    expect(screen.getByRole('status')).toHaveClass('spinner-border-sm');
   });
 });
